Add #include <string> to default lobster
My students love using lobster to visualize their code at Morehouse college intro to programing. The one issue is on the version of c++ they use the string library is included in iostream and on lobster it is not. They get errors when they try and use string. It would be great if their working code from elsewhere worked out of the box on lobster. (this doesn't fully solve the issue if they copy and paste their whole file with the includes, but that seems like it would be a bigger change to update the copy of iostream to include string in it the way it is in the most modern c++)